Can now retrieve Storable objects from database. database::utils::retrieve returns an std::optional containing a vector of Storable objects.
I decided that my decign for now will be to load up all of the Food
database objects into memory and use that rather than querying the
database constantly. Then updating the database as the vector grows.

Might need to change this design in the future.

Storable objects also now need to have a str() function to display
the database as the sqlite database would.

I don't like how I convert the data to string, then back to data.
Will need to change the desgin to use a std::variant rather than
have column data as a string.

I also realized that my Column structure is actually not representing
a column, but a row because we're traversing through a row
column by column.

Need to change this.